+++ wfx_fmac_driver/sl_wfx_version.h
@@ -18,9 +18,9 @@
 #define SL_WFX_VERSION_H
 
 /** X.x.x: Major version of the driver */
-#define FMAC_DRIVER_VERSION_MAJOR      2
+#define FMAC_DRIVER_VERSION_MAJOR      3
 /** x.X.x: Minor version of the driver */
-#define FMAC_DRIVER_VERSION_MINOR      4
+#define FMAC_DRIVER_VERSION_MINOR      0
 /** x.x.X: Revision of the driver */
 #define FMAC_DRIVER_VERSION_REVISION   0
 /** For release candidates, this is set to 1..254